Subclassed discord.Client
The main point was to create a message queue so the bot could iterate over messages slowly to avoid Discord's rate limit....
This turned out to be totally unecessary as discord.py handles this internally and is just letting you know in the stdout.
I think this was still a good subclass excersise though and may add things to myclient.Client in the future.

myclient.py
+ class Client(discord.Client)
  + self.outbound_messages is a deque for messages
DiscordBot.py
* client is now myclient.Client
commands.py
* client is now passed around rather than held in the global "client"
- removed some of the testing client.send_message() calls
config.py
  + RATE_LIMIT : for testing, to be removed